dev: use static webserver in container image
Here we update the example containerfile to use a static webserver,
rather than the dynamic webserver `npm start`. This is important to
evaluate how the routes will behave when hosted in production, on
Firebase static website hosting. We must ensure that all routes are
accessible out of the docroot directory.