ci: run dynamic server in container
We've been hosting the dapp as a static website, but the developers have
maintained it as a single-page application (SPA). Let's update the
containerfile to handle hosting the app so we can deploy via container
image. It's not ideal to use the webpack dev server for this use case,
but I'm prioritizing getting something up and running to unbreak the
extension for testnet users.