ci: perms clause for deploy
In addition to this diff, I made an out of band change, adding an IAM
policy binding for repository owner of "penumbra-zone" to the gcloud
service account. This allows us to use the github-actions service
account on more repos than just `penumbra-zone/penumbra`; we can now use
it on `penumbra-zone/*`.

Reorganizes the workflows so that the deploy workflow is triggered on
merge into main, and the deploy workflow calls the container workflow
and waits for its results.