bech32 test: working example cases
Updates the new bech32 test cases with working examples, showing how to
convert a base64-encoded Penumbra address to byte slice suitable for
bech32 re-encoding back to a string representation of the address.
Puzzlingly, the `bech32m` methods don't do what we want, so we use the
`bech32` methods instead.